+++ src/main/java/com/book/store/Service/CustomerService.java
@@ -32,7 +32,7 @@ private final UserRepository userRepository;
 
     // Delete customer
     @Transactional
-    public void deleteCustomer(Integer id) {
+    public void deleteCustomer(Long id) {
         customerRepository.deleteById(id);
     }
 
@@ -45,7 +45,7 @@ private final UserRepository userRepository;
     }
 
     // Find by ID
-    public CustomerApiDto findCustomerById(Integer id) {
+    public CustomerApiDto findCustomerById(Long id) {
         Customer customer = customerRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("Customer not found"));
         return customerMapper.toDTO(customer);
